tasks error fixed

+++ 102-counting_sort.c
@@ -28,7 +28,7 @@ void counting_sort(int *array, size_t size)
     count_array = malloc(sizeof(int) * (max + 1));
     if (!count_array)
         return;
-    for (i = 0; i <= (size_t)max; i++)
+    for (i = 0; i <= max; i++)
         count_array[i] = 0;
 
     /* Count occurrences of each element in the array */
@@ -36,7 +36,7 @@ void counting_sort(int *array, size_t size)
         count_array[array[i]]++;
 
     /* Update count_array to store the actual position of elements */
-    for (i = 1; i <= (size_t)max; i++)
+    for (i = 1; i <= max; i++)
         count_array[i] += count_array[i - 1];
 
     /* Create the sorted array */
@@ -60,7 +60,7 @@ void counting_sort(int *array, size_t size)
 
     /* Print the count array */
     printf("%d", count_array[0]);
-    for (i = 1; i <= (size_t)max; i++)
+    for (i = 1; i <= max; i++)
         printf(", %d", count_array[i]);
     printf("\n");
